chore: add test against clang compatibility
