+++ _tests/README.md
@@ -1,0 +1,36 @@
+# Compatibility Tests
+
+This directory contains the compatibility tests for the `go-argonize` package using Docker.
+
+Currently, we are testing the compatibility with the following language implementations:
+
+- Go vs PHP, Python and Clang
+
+Each test will generate a hashed password using Argon2id with random password,
+then compares/verifies the results against the ones generated by PHP, Python and Clang.
+
+## Usage
+
+To run the tests, run the blow command from the root directory of the repository:
+
+```bash
+$ docker compose build
+**snip**
+
+$ # NOTE: These tests will exit with status 0 if the tests pass, and 1 if they fail.
+
+$ # PHP vs Go
+$ docker compose --file ./.github/docker-compose.yml run --rm --remove-orphans php
+**snip**
+$ echo $?
+
+$ # Python vs Go
+$ docker compose --file ./.github/docker-compose.yml run --rm --remove-orphans python
+**snip**
+$ echo $?
+
+$ # Clang vs Go
+$ docker compose --file ./.github/docker-compose.yml run --rm --remove-orphans clang
+**snip**
+$ echo $?
+```
